♻️ 🏗️  Refactor file locations to align with FSD architecture
