+++ src/entities/forum/ui/PostCard.tsx
