+++ src/entities/index.ts
@@ -1,0 +1,2 @@
+
+export { default as PostCard } from "./forum/ui/PostCard";
