+++ src/features/index.ts
@@ -1,0 +1,3 @@
+export { fetchPosts } from './forum/api/posts';
+
+export { default as useForumPosts } from './forum/model/useForumPosts';
