+++ src/shared/index.ts
@@ -1,5 +1,3 @@
-import { CATEGORIES, FORUMBASEURL, PAGE_SIZE } from './const';
-import { Post } from './types';
+export { CATEGORIES, FORUM_BASE_URL, PAGE_SIZE } from './const';
+export type { Post } from './types';
 
-export { CATEGORIES, PAGE_SIZE, FORUMBASEURL };
-export type { Post };
