+++ src/widgets/index.ts
@@ -1,0 +1,3 @@
+export { default as Category } from "./forum/ui/Category";
+export { default as Nav } from "./forum/ui/Nav";
+export { default as PostList } from "./forum/ui/PostList";
